+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * `colnames` and `colnames<-` are now exported functions.
 * RStudio displays table in HTML viewer instead of markdown in the console.
+* `tt()` supports data.frame-like objects which also inherit from other classes, ex: `marginaleffects::slopes()`
 
 ## 0.2.0